Huge recalculation of attributes / damage / defense

+++ DiabloII.Items.Api/Items/Models/ItemProperty.cs
@@ -8,8 +8,8 @@ namespace DiabloII.Items.Api.Items.Responses
 		public string FormattedName { get; set; }
 		public string Name { get; set; }
 		public double Par { get; set; }
-        public int Minimum { get; set; }
-        public int Maximum { get; set; }
+        public double Minimum { get; set; }
+        public double Maximum { get; set; }
         public bool IsPercent { get; set; }
 		public string FirstChararacter { get; set; }
 		public double OrderIndex { get; set; }
