+++ DiabloII.Items.Reader/Records/WeaponRecord.cs
@@ -6,12 +6,11 @@
         public string Type { get; set; }
         public string Slot { get; set; }
 
-		public int MinimumOneHandedDamage { get; set; }
-		public int MaximumOneHandedDamage { get; set; }
-		public int MinimumTwoHandedDamage { get; set; }
-		public int MaximumTwoHandedDamage { get; set; }
-
-		public int StrengthRequired { get; set; }
-		public int DexterityRequired { get; set; }
+		public double MinimumOneHandedDamage { get; set; }
+		public double MaximumOneHandedDamage { get; set; }
+		public double MinimumTwoHandedDamage { get; set; }
+		public double MaximumTwoHandedDamage { get; set; }
+		public double StrengthRequired { get; set; }
+		public double DexterityRequired { get; set; }
 	}
 }
